fix: fixed issues with  keyword schema naming and writing inner classes that should have been their own class; the schema name is going to be correct more often from using x-schema-parser-id instead when appropriate; broke all args constructor for schemas that are arrays with this commit.

+++ hooks/post-process.js
@@ -77,7 +77,7 @@ function processSchema(generator, schemaName, schema, sourcePath, defaultJavaPac
   debugPostProcess(schema);
   const modelClass = applicationModel.getModelClass({schema: schema, schemaName: schemaName});
   const javaName = modelClass.getClassName();
-  if (schema.type() !== 'object' || _.startsWith(javaName, 'Anonymous')) {
+  if ((schema.type() && schema.type() !== 'object') || _.startsWith(javaName, 'Anonymous')) {
     debugPostProcess(`deleting ${filePath}`);
     fs.existsSync(filePath) && fs.unlinkSync(filePath);
   } else {
